requirements.txt
psycopg2-binary
python-dotenv
streamlit
google-generativeai